[FIX] to num 8 from issue list
