+++ minitodomal/src/componenets/AddTask.jsx
@@ -2,14 +2,23 @@ import '/src/styles/components/_btn.scss';
 import '/src/styles/components/_input.scss';
 import './AddTask.scss';
 
-function AddTask({ taskInput, handleInputChange, handleAddTask }) {
+function AddTask({ taskInput, handleChangeInput, tasks, setTaskInput, setTasks }) {
+  const handleAddTask = () => {
+    const newTask = {
+      id: tasks.length + 1,
+      text: taskInput,
+      done: false,
+    };
+    setTasks((prevTasks) => [...prevTasks, newTask]);
+    setTaskInput('');
+  };
   return (
     <div className={'add-task'}>
       <input
         className={'input-main'}
         type={'text'}
         value={taskInput}
-        onChange={handleInputChange}
+        onChange={handleChangeInput}
         placeholder={'enter task name'}
       />
       <button className={'btn-main'} onClick={handleAddTask}>
