+++ minitodomal/src/componenets/TaskList.scss
@@ -5,6 +5,7 @@
   height: 2rem;
   display: inline-block;
   cursor: pointer;
+
 }
 .delete{
   width: 2rem;
