+++ minitodomal/src/styles/style.scss
@@ -1,8 +1,5 @@
 @use 'abstracts';
-@use 'vendors';
-@use 'base';
 @use 'utilities';
 @use 'components';
-@use 'pages';
-@use 'themes';
-@use "layout";
+
+
